更改时间的有关的参数的类型为long

+++ src/main/java/com/jiefzz/ejoker/infrastructure/ProcessingMessageMailbox.java
@@ -78,7 +78,7 @@ public class ProcessingMessageMailbox<X extends IProcessingMessage<X, Y>, Y exte
 		}
 	}
 
-	public boolean isInactive(int timeoutSeconds) {
+	public boolean isInactive(long timeoutSeconds) {
 		return (System.currentTimeMillis() - lastActiveTime) >= timeoutSeconds;
 	}
